t5601: future-proof arguments to "case" statement
The wildcard argument to the expect_ssh() "case" statement was added
in 9f6976528b ("t5601: add more test cases for IPV6", 2015-02-21) but
was never used, not even in that commit.

Let's remove it. It's an anti-pattern to carry code we don't use, and
if this function ever does get passed four or five arguments we'd like
to die instead of silently dropping the new arguments.

While I'm at it fix another regression in that change. Before it
passing e.g. "blahblah" to expect_ssh() would cause the tests to fail,
but before it we'd only accept "none" to expect no output. Let's go
back to the stricter version of the function for clarity and to
future-proof it.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>