rev-parse: add error checking for --disambiguate
The --disambiguate option needs at least 4 characters of SHA-1 to do
its work, and at most 40, but if it was fed e.g. 3 or 41 characters it
would just print nothing and exit with 0.

Let's return an error instead, which makes it distinguishable from
e.g. giving it 10 characters that don't form the prefix for any SHA-1
object.

This test could be made portable to the SHA-256 backend if we exported
the hexdigit length for that hash algorithm to the tests. I'll leave
that to a future change, since likely we'll want to amend this and
multiple other tests to somehow test both hash versions.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>